requirements.txt
alembic==1.11.1
blinker==1.6.2
cachelib==0.9.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Deprecated==1.2.14
Flask==2.3.2
Flask-Caching==2.0.2
Flask-Limiter==3.3.1
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-QRcode==3.1.0
Flask-SQLAlchemy==3.0.5
idna==3.4
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
limits==3.5.0
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
ordered-set==4.1.0
packaging==23.1
Pillow==9.5.0
psycopg2-binary==2.9.6
Pygments==2.15.1
pypng==0.20220715.0
python-dotenv==1.0.0
qrcode==7.4.2
requests==2.31.0
rich==13.4.2
SQLAlchemy==2.0.16
typing_extensions==4.6.3
urllib3==2.0.3
Werkzeug==2.3.6
wrapt==1.15.0
